test without JSON::XS and add a compat note
JSON::XS 3.x is incompatible with JSON.pm < 2.90 in the exact way
that JSON::Any uses things in the test suite.

So we set the tests to ignore JSON::XS (using Test::Without::Module)
thanks to ether for suggesting it.

And we add a note about the situation to the POD so that others can
figure out what is going on.