fix the tests to properly exclude JSON::XS
A bug came up where JSON::XS and JSON.pm had incompatible revisions
but installed just fine. This was a problem because JSON.pm will silently
load JSON::XS in the background and caused explosions.

We now try to exclude JSON::XS except in the case we're actually testing it directly.

Also we've removed some of the tests for JSON::Syck as it's been deprecated. We also
add in tests for Cpanel::JSON::XS which got missed when it was added as a choice.